+++ Libraries/Fontawesome.php
@@ -24,10 +24,29 @@ class Fontawesome
 
             case '5.*':
             case '5.0.*':
+            case '5.0.8':
+            case '5.0.8.*':
+                $data = array(
+                    'href' => 'https://use.fontawesome.com/releases/v5.0.8/css/all.css',
+                    'integrity' => 'sha384-3AB7yXWz4OeoZcPbieVW64vVXEwADiYyAEhwilzWsLw+9FgqpyjjStpPnpBO8o8S',
+                    'crossorigin' => 'anonymous',
+                );
+
             case '5.0.6':
             case '5.0.6.*':
                 $data = array(
                     'href' => 'https://use.fontawesome.com/releases/v5.0.6/css/all.css',
+                    'integrity' => 'sha384-VY3F8aCQDLImi4L+tPX4XjtiJwXDwwyXNbkH7SHts0Jlo85t1R15MlXVBKLNx+dj',
+                    'crossorigin' => 'anonymous',
+                );
+                break;
+
+            case '5.0.4':
+            case '5.0.4.*':
+                $data = array(
+                    'href' => 'https://use.fontawesome.com/releases/v5.0.4/css/all.css',
+                    'integrity' => 'sha384-DmABxgPhJN5jlTwituIyzIUk6oqyzf3+XuP7q3VfcWA2unxgim7OSSZKKf0KSsnh',
+                    'crossorigin' => 'anonymous',
                 );
                 break;
 
@@ -49,10 +68,29 @@ class Fontawesome
 
             case '5.*':
             case '5.0.*':
+            case '5.0.8':
+            case '5.0.8.*':
+                $data = array(
+                    'src' => 'https://use.fontawesome.com/releases/v5.0.8/js/all.js',
+                    'integrity' => 'sha384-SlE991lGASHoBfWbelyBPLsUlwY1GwNDJo3jSJO04KZ33K2bwfV9YBauFfnzvynJ',
+                    'crossorigin' => 'anonymous',
+                );
+
             case '5.0.6':
             case '5.0.6.*':
                 $data = array(
                     'src' => 'https://use.fontawesome.com/releases/v5.0.6/js/all.js',
+                    'integrity' => 'sha384-0AJY8UERsBUKdWcyF3o2kisLKeIo6G4Tbd8Y6fbyw6qYmn4WBuqcvxokp8m2UzSD',
+                    'crossorigin' => 'anonymous',
+                );
+                break;
+
+            case '5.0.4':
+            case '5.0.4.*':
+                $data = array(
+                    'src' => 'https://use.fontawesome.com/releases/v5.0.4/js/all.js',
+                    'integrity' => 'sha384-nVi8MaibAtVMFZb4R1zHUW/DsTJpG/YwPknbGABVOgk5s6Vhopl6XQD/pTCG/DKB',
+                    'crossorigin' => 'anonymous',
                 );
                 break;