+++ Libraries/Fontawesomebase.php
@@ -12,7 +12,7 @@ namespace c975L\IncludeLibraryBundle\Libraries;
 /*
  * Fontawesome - https://fontawesome.com
  */
-class Fontawesomebrands
+class Fontawesome
 {
     /*
      * Refer to README.md for how-to add versions
@@ -27,16 +27,16 @@ class Fontawesomebrands
             case '5.0.8':
             case '5.0.8.*':
                 $data = array(
-                    'href' => 'https://use.fontawesome.com/releases/v5.0.8/css/brands.css',
-                    'integrity' => 'sha384-IiIL1/ODJBRTrDTFk/pW8j0DUI5/z9m1KYsTm/RjZTNV8RHLGZXkUDwgRRbbQ+Jh',
+                    'href' => 'https://use.fontawesome.com/releases/v5.0.8/css/fontawesome.css',
+                    'integrity' => 'sha384-q3jl8XQu1OpdLgGFvNRnPdj5VIlCvgsDQTQB6owSOHWlAurxul7f+JpUOVdAiJ5P',
                     'crossorigin' => 'anonymous',
                 );
 
             case '5.0.6':
             case '5.0.6.*':
                 $data = array(
-                    'href' => 'https://use.fontawesome.com/releases/v5.0.6/css/brands.css',
-                    'integrity' => 'sha384-rK0EPNdv8UCeRNPzX+96ARRlf9hZM+OukGceDTdbPH30DYcSI1x5QyBU7d2I2kHX',
+                    'href' => 'https://use.fontawesome.com/releases/v5.0.6/css/fontawesome.css',
+                    'integrity' => 'sha384-dbkYY2NmVwxaFrr4gq04oVh6w39ovmevsgD80Il1Od3hwpgREqyPb3XqbpaSwN4x',
                     'crossorigin' => 'anonymous',
                 );
                 break;
@@ -44,8 +44,8 @@ class Fontawesomebrands
             case '5.0.4':
             case '5.0.4.*':
                 $data = array(
-                    'href' => 'https://use.fontawesome.com/releases/v5.0.4/css/brands.css',
-                    'integrity' => 'sha384-1beec9tTZuu+KrTudmvRnGpK81r78DKCAXdphCvdG+PR+n/WCczsYPqTBTvYsM7z',
+                    'href' => 'https://use.fontawesome.com/releases/v5.0.4/css/fontawesome.css',
+                    'integrity' => 'sha384-xdTUmhbcetyLRVL4PAriRajOve+/5pjOiy5sJABnhXMcRMVc9HI9s2KmOCjjDK/P',
                     'crossorigin' => 'anonymous',
                 );
                 break;
@@ -71,16 +71,16 @@ class Fontawesomebrands
             case '5.0.8':
             case '5.0.8.*':
                 $data = array(
-                    'src' => 'https://use.fontawesome.com/releases/v5.0.8/js/brands.js',
-                    'integrity' => 'sha384-sCI3dTBIJuqT6AwL++zH7qL8ZdKaHpxU43dDt9SyOzimtQ9eyRhkG3B7KMl6AO19',
+                    'src' => 'https://use.fontawesome.com/releases/v5.0.8/js/fontawesome.js',
+                    'integrity' => 'sha384-7ox8Q2yzO/uWircfojVuCQOZl+ZZBg2D2J5nkpLqzH1HY0C1dHlTKIbpRz/LG23c',
                     'crossorigin' => 'anonymous',
                 );
 
             case '5.0.6':
             case '5.0.6.*':
                 $data = array(
-                    'src' => 'https://use.fontawesome.com/releases/v5.0.6/js/brands.js',
-                    'integrity' => 'sha384-4iSpDug9fizYiQRPpPafdAh5NaF8yzNMjOvu3veWgaFm0iIo8y4vUi7f3Yyz5WP1',
+                    'src' => 'https://use.fontawesome.com/releases/v5.0.6/js/fontawesome.js',
+                    'integrity' => 'sha384-rttr/ldR2uHigckjTCjMDe47ySeFVaL3Q7xUkJZir56u8Z8h/XnHJXHocgyfb25F',
                     'crossorigin' => 'anonymous',
                 );
                 break;
@@ -88,8 +88,8 @@ class Fontawesomebrands
             case '5.0.4':
             case '5.0.4.*':
                 $data = array(
-                    'src' => 'https://use.fontawesome.com/releases/v5.0.4/js/brands.js',
-                    'integrity' => 'sha384-dl3ONr32uA3YqpqKWzhXLs5k1YbKOn3dwiMbEP1S/XQMa3LPRwvJrhW7+lomL/uc',
+                    'src' => 'https://use.fontawesome.com/releases/v5.0.4/js/fontawesome.js',
+                    'integrity' => 'sha384-l7FyBM+wFIWpfmy8RYkWgEu/Me6Hrz98ijLu4nP3PkGbTtTCvtHB5ktI8hLEgEG3',
                     'crossorigin' => 'anonymous',
                 );
                 break;